Create README.md
I created a simple website that allows users to create and manage a to-do list. The website is built using HTML, CSS, and JavaScript. The main features of the website include the ability to create to-do items, edit to-do items, mark to-do items as complete, and delete to-do items. I faced a few challenges while building this project, such as learning how to use the different technologies involved. However, I was able to overcome these challenges by reading documentation and watching tutorials. I learned a lot from this project, such as how to use HTML, CSS, and JavaScript to build a functional website. I plan to continue working on this project and add new features, such as the ability to share to-do lists with other users.